#35 Move Prisma to Odyssey base
