+++ scylla-server/Dockerfile
@@ -16,11 +16,8 @@ RUN git init
 RUN git submodule add -f https://github.com/Northeastern-Electric-Racing/Odyssey-Base.git ./src/odyssey-base && \
     git submodule update --init --recursive
 
-RUN npx prisma@5.1.1 generate
-
-RUN npx prisma@5.1.1 migrate deploy
-
-RUN npm i --save-dev @types/ws
+RUN npm run prisma:generate && \
+    npm run prisma:migrate
 
 RUN npm run build
