--- scylla-server/src/prisma/schema.prisma
@@ -1,62 +1,0 @@
-generator client {
-    provider      = "prisma-client-js"
-    binaryTargets = ["native", "debian-openssl-1.1.x", "linux-arm64-openssl-1.1.x"]
-}
-
-datasource db {
-    provider = "sqlite"
-    url      = "file:./mydatabase.db"
-}
-
-model Run {
-    id           Int      @id @unique @default(autoincrement())
-    locationName String
-    location     Location @relation(fields: [locationName], references: [name])
-    driverId     String?
-    driver       Driver?  @relation(fields: [driverId], references: [username])
-    systemId     String?
-    system       System?  @relation(fields: [systemId], references: [name])
-    data         Data[]
-    time         DateTime
-}
-
-model Location {
-    name      String @id @unique
-    latitude  Float
-    longitude Float
-    radius    Float
-    runs      Run[]
-}
-
-model Driver {
-    username String @id @unique
-    runs     Run[]
-}
-
-model System {
-    name String @id @unique
-    runs Run[]
-}
-
-model Data {
-    id           Int      @id @unique @default(autoincrement())
-    value        Int
-    dataTypeName String
-    dataType     DataType @relation(fields: [dataTypeName], references: [name])
-    time         Int
-    runId        Int
-    run          Run      @relation(fields: [runId], references: [id])
-}
-
-model DataType {
-    name     String @id @unique
-    unit     String
-    data     Data[]
-    node     Node   @relation(fields: [nodeName], references: [name])
-    nodeName String
-}
-
-model Node {
-    name      String     @id @unique
-    dataTypes DataType[]
-}
